refactor(#115): 패키지 구조 변경
- consumer 위치 application level로 변경
