+++ src/main/java/com/mos/backend/studymaterials/application/consumer/FileUploadConsumer.java
@@ -1,8 +1,8 @@
-package com.mos.backend.studymaterials.infrastructure.event.consumer;
+package com.mos.backend.studymaterials.application.consumer;
 
 import com.mos.backend.common.event.Event;
 import com.mos.backend.studymaterials.application.StudyMaterialService;
-import com.mos.backend.studymaterials.application.event.FileUploadFailedEventPayload;
+import com.mos.backend.studymaterials.application.event.FileUploadFailedEventPayloadWithNotification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -14,8 +14,8 @@ public class FileUploadConsumer {
     private final StudyMaterialService studyMaterialService;
 
     @EventListener
-    public void handleFileUploadFailedEvent(Event<FileUploadFailedEventPayload> event) {
-        FileUploadFailedEventPayload payload = event.getPayload();
+    public void handleFileUploadFailedEvent(Event<FileUploadFailedEventPayloadWithNotification> event) {
+        FileUploadFailedEventPayloadWithNotification payload = event.getPayload();
         studyMaterialService.delete(payload.getFilePath());
     }
 }
