+++ src/main/java/com/mos/backend/studymembers/application/consumer/StudyMemberConsumer.java
@@ -1,8 +1,8 @@
-package com.mos.backend.studyrequirements.infrastructure.event.consumer;
+package com.mos.backend.studymembers.application.consumer;
 
 import com.mos.backend.common.event.Event;
 import com.mos.backend.studies.application.event.StudyCreatedEventPayload;
-import com.mos.backend.studyrequirements.application.StudyRequirementService;
+import com.mos.backend.studymembers.application.StudyMemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -13,13 +13,13 @@ import static org.springframework.transaction.event.TransactionPhase.BEFORE_COMM
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class StudyRequirementConsumer {
+public class StudyMemberConsumer {
 
-    private final StudyRequirementService studyRequirementService;
+    private final StudyMemberService studyMemberService;
 
     @TransactionalEventListener(phase = BEFORE_COMMIT)
     public void handleStudyCreatedEvent(Event<StudyCreatedEventPayload> event) {
         StudyCreatedEventPayload payload = event.getPayload();
-        studyRequirementService.createOrUpdateOrDelete(payload.getStudyId(), payload.getRequestDto().getRequirements());
+        studyMemberService.createStudyLeader(payload.getStudyId(), payload.getUserId());
     }
 }
